Basic implementation of posts is working.
Posts can now be created, searched by ID, and gotten in a
batch of all of them. The successfully save to the database.
Next to be done would be to determine what other meta-data
goes with a post, add any attachments (videos, images) to the
post, and associate them with a 'user' when the post is made.

This means that a front end for all of this will need to be made
in the coming days.